requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
boto3==1.28.85
Werkzeug==2.3.7
python-dotenv==1.0.0